refactor(kv-store): foyer-style builder and module layout
Builder, after studying foyer's HybridCacheBuilder:
- All knobs are chainable setters (with_requery_interval / with_resolve_deadline
  / with_cacheable_pin_percent); the public KvStoreConfig struct is gone — no
  options struct to churn when a knob is added.
- Optional components are separate methods, not Option parameters:
  rank(rank, pool) declares the tier-less mode behind the same API (foyer's
  storage-disabled hybrid mode), rank_with_tier(rank, pool, tier) the full one.
- Derived values (per-rank cacheable pin budgets) are computed at build(), so
  setter order can never skew them — the previous builder computed the budget
  inside .rank(), silently wrong if the percent was set afterwards.

Module layout: lib.rs is interface surface only (crate doc, module decls,
re-exports — including KvPrefix from the engine contract so callers have one
import point). Implementation splits into builder.rs (construction and the
freeze), store.rs (KvStore + RankState + resolve/seal/retire), policy.rs (the
caller vocabulary: CancelProbe, CacheScope, ResolvePolicy, SaveClass,
SaveCursor). tier.rs / stats.rs / testkit.rs unchanged.

Also: KvPrefix's Debug now includes the rank field; RankDecl type alias.
